:package: Refactor (move) `PythonTestMixIn` into `tests.mixin`
