+++ testapps/on_device_unit_tests/test_app/tests/test_requirements.py
@@ -1,6 +1,6 @@
 
-from main import PythonTestMixIn
 from unittest import TestCase
+from .mixin import PythonTestMixIn
 
 
 class NumpyTestCase(PythonTestMixIn, TestCase):
